Fixed navbar + updated examples page
Fixed some things so the navbar look is consistent across all pages and now has a logo. Had to update all html files for the logo. Currently there is an error with the quiz page navbar.

Added content to consistency and shortcuts section of examples page. Also added slideshows to show different pages of websites.